Improve platform-independent bit_c[lt]z32
I don't like `+ 1` before multiplication in bit_clz32, but I don't see a
clear way to get rid of that either. This would require something other
than a de Brujin sequence, and constructing it didn't prove to be easy.
Regardless, one cycle of latency shouldn't be significant given all the
shifts and ORs.