Implement 64-bit bswap with D&C
This is a little futile, as most modern compilers detect byte-swap
patterns anyway. However, compiling this on something that doesn't, e.g.
old MSVC, shows that the new code slightly improves performance. The
impact would be larger on non-super-scalar processors, but can be
reproduced even on x86-64.

The effect is a lot less pronounced on 32-bit bswap and may even reduce
performance on superscalar processors, so we don't implement it that way.