tap_user: Always send DHCP packets to port 68
The SerenityOS DHCPClient sends DHCP packets from a random port and
listens on port 68, so the response would get lost.

The DHCP RFC says that that "DHCP messages from a server to a client
are sent to the 'DHCP client' port (68)."
(https://datatracker.ietf.org/doc/html/rfc1541#section-4.1)
So SerenityOS' behaviour should be valid.